Bugfix and start work on price peak planner

+++ custom_components/energy_planner/planner/basic_planner.py
@@ -29,12 +29,12 @@ async def plan_day(hass: HomeAssistant, nordpool_values: [dict], config: dict):
         x for x in nordpool_values if config["earliest_discharge"] <= x["start"]
     ]
     cheapest_hours = sorted(
-        sorted(charge_hours, key=lambda x: x["value"])[: config["nr_of_charge_hours"] * 4],
+        sorted(charge_hours, key=lambda x: x["value"])[: int(config["nr_of_charge_hours"] * 4)],
         key=lambda x: x["start"],
     )
     expensive_hours = sorted(
         sorted(discharge_hours, key=lambda x: x["value"], reverse=True)[
-            : config["nr_of_discharge_hours"] * 4
+            : int(config["nr_of_discharge_hours"] * 4)
         ],
         key=lambda x: x["start"],
     )
